+++ src/libraries/DisputeLib.sol
@@ -16,25 +16,14 @@ import {
     OriginatorReportStatus
 } from "src/Types.sol";
 
-/// @title DisputeLib
-/// @notice Deployed library for dispute and originator report operations.
-/// @dev Called via DELEGATECALL from SapienCore; operates on the caller's ERC-7201 storage.
-///      Uses ReputationLib (also via DELEGATECALL) for reputation updates.
 library DisputeLib {
-    // keccak256(abi.encode(uint256(keccak256("sapien.storage.SapienCore")) - 1)) & ~bytes32(uint256(0xff))
     function _getStorage() private pure returns (EngineStorage storage $) {
         assembly {
             $.slot := 0xb21037e32bd67da4126ec23c3d75228183c819f055709f5aa59aa33cc3fd2b00
         }
     }
 
-    // ════════════════════════════════════════════════════════════════════
-    // Disputes — Consensus Outcome Challenges
-    // ════════════════════════════════════════════════════════════════════
-
-    /// @notice Open a dispute on a consensus outcome during the challenge period.
     function openDispute(bytes32 projectId, uint256 index, bytes32 evidenceHash, string calldata evidenceCid) public {
-        // SEC-L-03: Require non-empty evidence hash
         if (evidenceHash == bytes32(0)) revert ISapienCore.InvalidEvidenceHash();
 
         EngineStorage storage $ = _getStorage();
@@ -48,10 +37,8 @@ library DisputeLib {
         if (contrib.challengeEndsAt == 0) revert ISapienCore.ConsensusNotComputed();
         if (block.timestamp > contrib.challengeEndsAt) revert ISapienCore.DisputeWindowClosed();
 
-        // SEC-C-01: disputes keyed by nonce to prevent cross-nonce poisoning
         Dispute storage dispute = $.disputes[projectId][index][nonce];
 
-        // SEC-H-03: only one dispute per (projectId, index, nonce) — block reopening
         if (dispute.status == DisputeStatus.Open) revert ISapienCore.DisputeAlreadyOpen();
         if (dispute.status == DisputeStatus.Rejected || dispute.status == DisputeStatus.Upheld) {
             revert ISapienCore.DisputeAlreadyClosed();
@@ -66,19 +53,18 @@ library DisputeLib {
         $.vault.lockContributor(msg.sender, bondAmount);
 
         dispute.challenger = msg.sender;
-        dispute.openedAt = block.timestamp;
+        dispute.openedAt = uint48(block.timestamp);
         dispute.status = DisputeStatus.Open;
         dispute.bondAmount = bondAmount;
         dispute.evidenceHash = evidenceHash;
 
         if (contrib.status == ContributionStatus.Accepted) {
-            contrib.challengeEndsAt = block.timestamp + C.DISPUTE_RESOLUTION_DEADLINE;
+            contrib.challengeEndsAt = uint48(block.timestamp + C.DISPUTE_RESOLUTION_DEADLINE);
         }
 
         emit ISapienCore.DisputeOpened(projectId, index, msg.sender, bondAmount, evidenceCid);
     }
 
-    /// @notice Execute the "upheld" path for a dispute (used by both resolve and escalate).
     function upholdDispute(bytes32 projectId, uint256 index, uint256 nonce) public {
         EngineStorage storage $ = _getStorage();
         Dispute storage dispute = $.disputes[projectId][index][nonce];
@@ -87,7 +73,7 @@ library DisputeLib {
         address rewardToken = proj.rewardToken;
 
         dispute.status = DisputeStatus.Upheld;
-        dispute.resolvedAt = block.timestamp;
+        dispute.resolvedAt = uint40(block.timestamp);
         $.vault.unlockContributor(dispute.challenger, dispute.bondAmount);
 
         emit ISapienCore.DisputeResolved(projectId, index, true);
@@ -100,13 +86,10 @@ library DisputeLib {
             }
             ReputationLib.update(contrib.contributor, proj.requiredSkill, false, 0);
 
-            // Mark this accepted contribution as lifecycle-complete from the pipeline
-            // perspective so upheld disputes do not deadlock project completion.
             if ($.pendingContributions[projectId] > 0) {
                 $.pendingContributions[projectId]--;
             }
         } else if (contrib.status == ContributionStatus.Rejected) {
-            // SEC-H-04: Cap total payout to rewardRate (contributor + challenger share a single budget)
             uint256 maxPayout = contrib.rewardRate;
             uint256 challengerReward = (maxPayout * C.DISPUTE_CHALLENGER_REWARD_BPS) / C.BPS;
             uint256 compensation = maxPayout - challengerReward;
@@ -125,30 +108,23 @@ library DisputeLib {
         }
     }
 
-    /// @notice Execute the "rejected" path for a dispute.
     function rejectDispute(bytes32 projectId, uint256 index, uint256 nonce) public {
         EngineStorage storage $ = _getStorage();
         Dispute storage dispute = $.disputes[projectId][index][nonce];
         Contribution storage contrib = $.contributions[projectId][index];
 
         dispute.status = DisputeStatus.Rejected;
-        dispute.resolvedAt = block.timestamp;
+        dispute.resolvedAt = uint40(block.timestamp);
         $.vault.slashContributor(dispute.challenger, dispute.bondAmount);
 
         if (contrib.status == ContributionStatus.Accepted) {
-            contrib.challengeEndsAt = block.timestamp;
+            contrib.challengeEndsAt = uint48(block.timestamp);
         }
 
         emit ISapienCore.DisputeResolved(projectId, index, false);
     }
 
-    // ════════════════════════════════════════════════════════════════════
-    // Originator Accountability
-    // ════════════════════════════════════════════════════════════════════
-
-    /// @notice Report an originator for misconduct.
     function reportOriginator(bytes32 projectId, bytes32 evidenceHash) public {
-        // SEC-L-03: Require non-empty evidence hash
         if (evidenceHash == bytes32(0)) revert ISapienCore.InvalidEvidenceHash();
 
         EngineStorage storage $ = _getStorage();
@@ -169,7 +145,7 @@ library DisputeLib {
         $.vault.lockContributor(msg.sender, bondAmount);
 
         report.reporter = msg.sender;
-        report.reportedAt = block.timestamp;
+        report.reportedAt = uint48(block.timestamp);
         report.status = OriginatorReportStatus.Open;
         report.bondAmount = bondAmount;
         report.evidenceHash = evidenceHash;
@@ -177,15 +153,13 @@ library DisputeLib {
         emit ISapienCore.OriginatorReported(projectId, msg.sender, bondAmount);
     }
 
-    /// @notice Execute the "upheld" path for an originator report.
-    /// @param includeReporterReward If true, give reporter a reward from escrow (used by resolve, not escalate)
     function upholdOriginatorReport(bytes32 projectId, bool includeReporterReward) public {
         EngineStorage storage $ = _getStorage();
         OriginatorReport storage report = $.originatorReports[projectId];
         Project storage proj = $.projects[projectId];
 
         report.status = OriginatorReportStatus.Upheld;
-        report.resolvedAt = block.timestamp;
+        report.resolvedAt = uint40(block.timestamp);
         $.vault.unlockContributor(report.reporter, report.bondAmount);
 
         uint256 originatorStake = $.originatorLockedStake[projectId];
@@ -205,17 +179,16 @@ library DisputeLib {
 
         ReputationLib.update(proj.originator, C.ORIGINATOR_ROLE_KEY, false, 0);
         proj.status = ProjectStatus.Cancelled;
-        proj.cancelledAt = block.timestamp;
+        proj.cancelledAt = uint48(block.timestamp);
         emit ISapienCore.ProjectCancelled(projectId);
     }
 
-    /// @notice Execute the "rejected" path for an originator report.
     function rejectOriginatorReport(bytes32 projectId) public {
         EngineStorage storage $ = _getStorage();
         OriginatorReport storage report = $.originatorReports[projectId];
 
         report.status = OriginatorReportStatus.Rejected;
-        report.resolvedAt = block.timestamp;
+        report.resolvedAt = uint40(block.timestamp);
         $.vault.slashContributor(report.reporter, report.bondAmount);
 
         emit ISapienCore.OriginatorReportResolved(projectId, false);
