+++ src/libraries/ReputationLib.sol
@@ -6,23 +6,17 @@ import {ISapienCore} from "src/interfaces/ISapienCore.sol";
 import {EngineStorage, Reputation} from "src/Types.sol";
 import {Math} from "@openzeppelin/contracts/utils/math/Math.sol";
 
-/// @title ReputationLib
-/// @notice Deployed library for reputation system operations.
-/// @dev Called via DELEGATECALL from SapienCore; operates on the caller's ERC-7201 storage.
 library ReputationLib {
-    // keccak256(abi.encode(uint256(keccak256("sapien.storage.SapienCore")) - 1)) & ~bytes32(uint256(0xff))
     function _getStorage() private pure returns (EngineStorage storage $) {
         assembly {
             $.slot := 0xb21037e32bd67da4126ec23c3d75228183c819f055709f5aa59aa33cc3fd2b00
         }
     }
 
-    /// @notice Get reputation score with lazy decay applied
     function getScore(address user, bytes32 role) public view returns (uint256) {
         return getScoreCached(user, role, _getStorage().decayRateBps);
     }
 
-    /// @notice Get reputation score with pre-cached decayRateBps (avoids redundant SLOAD in loops)
     function getScoreCached(address user, bytes32 role, uint256 cachedDecayBps) public view returns (uint256) {
         Reputation storage rep = _getStorage().reputation[user][role];
         if (rep.lastUpdated == 0) return C.DEFAULT_REPUTATION;
@@ -30,7 +24,7 @@ library ReputationLib {
         uint256 score = rep.score;
 
         if (cachedDecayBps > 0) {
-            uint256 elapsed = block.timestamp - rep.lastUpdated;
+            uint256 elapsed = block.timestamp - uint256(rep.lastUpdated);
             if (elapsed >= 1 days) {
                 uint256 decayAmount = Math.mulDiv(score * cachedDecayBps, elapsed / 1 days, C.BPS);
                 score = score > decayAmount + C.MIN_REPUTATION ? score - decayAmount : C.MIN_REPUTATION;
@@ -40,22 +34,21 @@ library ReputationLib {
         return score;
     }
 
-    /// @notice Update reputation after an action (success or failure with optional bonus)
     function update(address user, bytes32 skillId, bool success, uint256 bonus) public {
         EngineStorage storage $ = _getStorage();
         Reputation storage rep = $.reputation[user][skillId];
 
         if (rep.lastUpdated == 0) {
-            rep.score = C.DEFAULT_REPUTATION;
-            rep.lastUpdated = block.timestamp;
+            rep.score = uint16(C.DEFAULT_REPUTATION);
+            rep.lastUpdated = uint48(block.timestamp);
         }
 
         uint256 currentScore = rep.score;
 
         if ($.decayRateBps > 0) {
-            uint256 elapsed = block.timestamp - rep.lastUpdated;
+            uint256 elapsed = block.timestamp - uint256(rep.lastUpdated);
             if (elapsed >= 1 days) {
-                uint256 decayAmount = Math.mulDiv(currentScore * $.decayRateBps, elapsed / 1 days, C.BPS);
+                uint256 decayAmount = Math.mulDiv(currentScore * uint256($.decayRateBps), elapsed / 1 days, C.BPS);
                 currentScore =
                     currentScore > decayAmount + C.MIN_REPUTATION ? currentScore - decayAmount : C.MIN_REPUTATION;
             }
@@ -71,7 +64,7 @@ library ReputationLib {
             uint256 today = block.timestamp / 1 days;
             if (rep.dailyGainDate != today) {
                 rep.dailyGain = 0;
-                rep.dailyGainDate = today;
+                rep.dailyGainDate = uint32(today);
             }
 
             uint256 gain = C.SUCCESS_INCREASE + bonus;
@@ -82,15 +75,15 @@ library ReputationLib {
 
             if (gain > 0) {
                 currentScore = currentScore + gain > C.MAX_REPUTATION ? C.MAX_REPUTATION : currentScore + gain;
-                rep.dailyGain += gain;
+                rep.dailyGain += uint16(gain);
             }
         } else {
             uint256 penalty = C.REJECTION_DECREASE;
             currentScore = currentScore > penalty + C.MIN_REPUTATION ? currentScore - penalty : C.MIN_REPUTATION;
         }
 
-        rep.score = currentScore;
-        rep.lastUpdated = block.timestamp;
+        rep.score = uint16(currentScore);
+        rep.lastUpdated = uint48(block.timestamp);
 
         if (currentScore != oldScore) {
             emit ISapienCore.ReputationUpdated(user, skillId, oldScore, currentScore);
